Get data for an athlete

Get data for an athlete by web scraping, does not require authentication.

Usage

athl_fun(athl_num, trace = TRUE)

Arguments

athl_num

numeric vector of athlete id(s) used by Strava, as character string

trace

logical indicating if output is returned to console

Value

A list for each athlete, where each element is an additional list with elements for the athlete's information. The list elements are named using the athlete id numbers.

Details

The athlete id is assigned to the user during registration with Strava and this must be known to use the function. Some users may have privacy settings that prevent public access to account information (a message indicating as such will be returned by the function). The function scrapes data using the following URL with the appended athlete id, e.g., https://www.strava.com/athletes/2837007. Opening the URL in a web browser can verify if the data can be scraped. Logging in to the Strava account on the website may also be required before using this function.

Examples

## single athlete
athl_fun('2837007')
#> 2837007 1 of 1 
#> $`2837007`
#> $`2837007`$name
#> [1] "Marcus Beck"
#> 
#> $`2837007`$location
#> [1] "United States"
#> 
#> $`2837007`$follow
#>   Followers Following
#> 1        83        83
#> 
#> $`2837007`$monthly
#>   Distance Moving Time
#> 1  71.3 mi     7:20:30
#> 
#> $`2837007`$recent
#>            Date           Name Distance Elevation  Time
#> 1     Yesterday    Morning Run   5.1 km       5 m 29:54
#> 2 July 13, 2025    Evening Run   8.1 km       5 m 45:21
#> 3 July 11, 2025 Afternoon Ride   4.7 km       9 m 12:49
#> 
#> 

## multiple athletes
athl_fun(c('2837007', '2527465'))
#> 2837007 1 of 2 
#> 2527465 2 of 2 
#> $`2837007`
#> $`2837007`$name
#> [1] "Marcus Beck"
#> 
#> $`2837007`$location
#> [1] "United States"
#> 
#> $`2837007`$follow
#>   Followers Following
#> 1        83        83
#> 
#> $`2837007`$monthly
#>   Distance Moving Time
#> 1  71.3 mi     7:20:30
#> 
#> $`2837007`$recent
#>            Date           Name Distance Elevation  Time
#> 1     Yesterday    Morning Run   5.1 km       5 m 29:54
#> 2 July 13, 2025    Evening Run   8.1 km       5 m 45:21
#> 3 July 11, 2025 Afternoon Ride   4.7 km       9 m 12:49
#> 
#> 
#> $`2527465`
#> $`2527465`$name
#> [1] "Pedro Villarroel"
#> 
#> $`2527465`$location
#> [1] "Argentina"
#> 
#> $`2527465`$follow
#>   Followers Following
#> 1       764       342
#> 
#> $`2527465`$monthly
#>   Distance Moving Time
#> 1 415.9 km    17:31:46
#> 
#> $`2527465`$recent
#>            Date                    Name Distance Elevation    Time
#> 1 July 13, 2025  3hr Z2 en el Sarmiento 116.8 km     543 m 4:28:48
#> 2 July 12, 2025                 Regreso  20.8 km      44 m 1:03:15
#> 3 July 12, 2025 Lucila + Ida a Martinez  21.9 km      69 m 1:03:07
#> 
#>