Get speed splits in a dataframe

Allows the return of speed splits of multiple rides.

Usage

get_spdsplits(act_id, stoken, units = "metric")

Arguments

act_id

a vector of activity IDs. These are easily found in the data.frame returned by compile_activities

stoken

A config object created using the strava_oauth function

units

chr string indicating plot units as either metric or imperial

Value

a data frame containing the splits of the activity or activities selected.

Author

Marcus Beck

Examples

if (FALSE) { # \dontrun{
# get my activities
stoken <- httr::config(token = strava_oauth(app_name, app_client_id, app_secret, cache = TRUE))
my_acts <- get_activity_list(stoken)

# compile activities
acts_data <- compile_activities(my_acts)

# get spdsplits for all activities
spd_splits <- purrr::map_df(acts_data$id, get_spdsplits, stoken = stoken, 
     units = 'metric', .id = 'id')
} # }