Plot speed by splits

Plot average speed by splits for a single activity

Usage

plot_spdsplits(act_data, ...)

# S3 method for class 'list'
plot_spdsplits(
  act_data,
  stoken,
  acts = 1,
  id = NULL,
  units = "metric",
  fill = "darkblue",
  ...
)

# Default S3 method
plot_spdsplits(act_data, stoken, units = "metric", fill = "darkblue", ...)

Arguments

act_data

an activities list object returned by get_activity_list or a data.frame returned by compile_activities

...

arguments passed to other methods

stoken

A config object created using the strava_oauth function

acts

numeric indicating which activity to plot based on index in the activities list, defaults to most recent

id

optional character vector to specify the id(s) of the activity/activities to plot, acts is ignored if provided

units

chr string indicating plot units as either metric or imperial

fill

chr string of fill color for profile

Value

plot of average distance for each split value in the activity

Details

The average speed per split is plotted, including a dashed line for the overall average. The final split is typically not a complete km or mile.

Author

Marcus Beck

Examples

if (FALSE) { # \dontrun{
# get my activities
stoken <- httr::config(token = strava_oauth(app_name, app_client_id, app_secret, cache = TRUE))
my_acts <- get_activity_list(stoken)

# default
plot_spdsplits(my_acts, stoken, acts = 1)
} # }