Get data for an athlete

Get data for an athlete by web scraping, does not require authentication.

athl_fun(athl_num, trace = TRUE)

Arguments

athl_num
numeric vector of athlete id(s) used by Strava

trace
logical indicating if output is returned to console

Value

A list for each athlete, where each element is an additional list with elements for the athlete name, location, units of measurement, monthly data, recent activities, and achievements. The list elements are named using the athlete id numbers. NA will be returned if the data for an athlete could not be accessed.

Details

The athlete id is assigned to the user during registration with Strava and this must be known to use the function. Some users may have privacy settings that prevent public access to account information (a message indicating as such will be returned by the function). The function scrapes data from https://www.strava.com/athletes/ with the appended athlete id, e.g., https://www.strava.com/athletes/2837007. Opening the URL in a web browser can verify if the data can be scraped. Logging in to the Strava account on the website may also be required before using this function.

Examples

## single athlete
athl_fun(2837007)

#> 2837007 1 of 1 

#> $`2837007`
#> $`2837007`$name
#> [1] "Marcus Beck"
#> 
#> $`2837007`$location
#> [1] "Irvine, California"
#> 
#> $`2837007`$units
#> [1] "mi" "ft"
#> 
#> $`2837007`$monthly
#>        month     miles hours elev_gain
#> 1 2019-07-01 273.54686    20       893
#> 2 2019-08-01 342.39727    25      1126
#> 3 2019-09-01 284.57744    20      1041
#> 4 2019-10-01  48.97586     7       773
#> 5 2019-11-01 165.17040    14       545
#> 6 2019-12-01 124.46997    11       553
#> 7 2020-01-01   0.00000     0         0
#> 
#> $`2837007`$recent
#>           id          name type startDateLocal distance elevation movingTime
#> 1 2960494076     Lunch Run  run     2019-12-27      3.2       158      24:51
#> 2 2958411997 Afternoon Run  run     2019-12-26      3.2       160      24:57
#> 3 2955825575   Morning Run  run     2019-12-25      3.2       161      24:50
#> 
#> $`2837007`$achievements
#>                         description             timeago
#> 1             PR on Snell Bridge up 2019-12-15 18:04:34
#> 2               PR on 21-9st Sprint 2019-12-05 15:09:05
#> 3 2nd fastest time on 21-9st Sprint 2019-12-10 11:44:20
#> 4 2nd fastest time on 9-21st Sprint 2019-12-05 21:40:46
#> 
#> 

## multiple athletes
athl_fun(c(2837007, 2527465))

#> 2837007 1 of 2 
#> 2527465 2 of 2 

#> $`2837007`
#> $`2837007`$name
#> [1] "Marcus Beck"
#> 
#> $`2837007`$location
#> [1] "Irvine, California"
#> 
#> $`2837007`$units
#> [1] "mi" "ft"
#> 
#> $`2837007`$monthly
#>        month     miles hours elev_gain
#> 1 2019-07-01 273.54686    20       893
#> 2 2019-08-01 342.39727    25      1126
#> 3 2019-09-01 284.57744    20      1041
#> 4 2019-10-01  48.97586     7       773
#> 5 2019-11-01 165.17040    14       545
#> 6 2019-12-01 124.46997    11       553
#> 7 2020-01-01   0.00000     0         0
#> 
#> $`2837007`$recent
#>           id          name type startDateLocal distance elevation movingTime
#> 1 2960494076     Lunch Run  run     2019-12-27      3.2       158      24:51
#> 2 2958411997 Afternoon Run  run     2019-12-26      3.2       160      24:57
#> 3 2955825575   Morning Run  run     2019-12-25      3.2       161      24:50
#> 
#> $`2837007`$achievements
#>                         description             timeago
#> 1             PR on Snell Bridge up 2019-12-15 18:04:34
#> 2               PR on 21-9st Sprint 2019-12-05 15:09:05
#> 3 2nd fastest time on 21-9st Sprint 2019-12-10 11:44:20
#> 4 2nd fastest time on 9-21st Sprint 2019-12-05 21:40:46
#> 
#> 
#> $`2527465`
#> $`2527465`$name
#> [1] "Pedro Villarroel"
#> 
#> $`2527465`$location
#> [1] "Buenos Aires, Ciudad Autónoma de Buenos Aires, Argentina"
#> 
#> $`2527465`$units
#> [1] "km" "m" 
#> 
#> $`2527465`$monthly
#>        month  miles hours elev_gain
#> 1 2019-07-01 95.858     5      1057
#> 2 2019-08-01  6.859     0        54
#> 3 2019-09-01  0.000     0         0
#> 4 2019-10-01 76.962     4       877
#> 5 2019-11-01 67.650     4       960
#> 6 2019-12-01  3.638     0        28
#> 7 2020-01-01 34.464     1       650
#> 
#> $`2527465`$recent
#>           id                name type startDateLocal distance elevation
#> 1 2974198613   Comenzando el año ride     2020-01-01     34.4       650
#> 2 2918189135 Biking Buenos Aires ride     2019-12-07      3.6        28
#>   movingTime
#> 1    1:43:25
#> 2      10:52
#> 
#> $`2527465`$achievements
#>                                                description             timeago
#> 1                     PR on Hipódromo - Palermo a Belgrano 2020-01-01 21:02:12
#> 2          PR on Bicisenda Zoologico (santa fe-Libertador) 2020-01-01 21:02:12
#> 3               PR on Las Heras (Pueyrredón - Pza. Italia) 2020-01-01 21:02:12
#> 4 2nd fastest time on SANTA FE INTEGRAL - #teamsilentcycle 2020-01-01 21:02:12
#> 5                     3rd fastest time on Plaza San Martín 2020-01-01 21:02:12
#> 
#>