Get data for an athlete

Get data for an athlete by web scraping, does not require authentication.

athl_fun(athl_num, trace = TRUE)

Arguments

athl_num
numeric vector of athlete id(s) used by Strava

trace
logical indicating if output is returned to console

Value

A list for each athlete, where each element is an additional list with elements for the athlete name, location, units of measurement, monthly data, recent activities, and achievements. The list elements are named using the athlete id numbers. NA will be returned if the data for an athlete could not be accessed.

Details

The athlete id is assigned to the user during registration with Strava and this must be known to use the function. Some users may have privacy settings that prevent public access to account information (a message indicating as such will be returned by the function). The function scrapes data from https://www.strava.com/athletes/ with the appended athlete id, e.g., https://www.strava.com/athletes/2837007. Opening the URL in a web browser can verify if the data can be scraped. Logging in to the Strava account on the website may also be required before using this function.

Examples

## single athlete
athl_fun(2837007)

#> 2837007 1 of 1 

#> $`2837007`
#> $`2837007`$name
#> [1] "Marcus Beck"
#> 
#> $`2837007`$location
#> [1] "Irvine, California"
#> 
#> $`2837007`$units
#> [1] "mi" "ft"
#> 
#> $`2837007`$monthly
#>        month     miles hours elev_gain
#> 1 2019-05-01 351.16172    25      1266
#> 2 2019-06-01 283.40678    21      1132
#> 3 2019-07-01 273.54686    20       893
#> 4 2019-08-01 342.39727    25      1126
#> 5 2019-09-01 284.57744    20      1041
#> 6 2019-10-01  48.97586     7       773
#> 7 2019-11-01  56.55658     4       147
#> 
#> $`2837007`$recent
#>           id           name type startDateLocal distance elevation movingTime
#> 1 2859540960   Morning Ride ride     2019-11-11      3.5        18      16:29
#> 2 2858493438 Afternoon Ride ride     2019-11-11     36.2       339    2:21:55
#> 3 2858170842    Morning Run  run     2019-11-08      3.2        47      23:36
#> 
#> $`2837007`$achievements
#>                          description             timeago
#> 1   2nd best estimated 1 mile effort 2019-11-09 12:01:09
#> 2       3rd best estimated 1k effort 2019-11-09 12:01:09
#> 3 3rd best estimated 1/2 mile effort 2019-11-09 12:01:09
#> 4     3rd best estimated 400m effort 2019-11-09 12:01:09
#> 5  2nd fastest time on 21-9st Sprint 2019-11-12 12:10:31
#> 
#> 

## multiple athletes
athl_fun(c(2837007, 2527465))

#> 2837007 1 of 2 
#> 2527465 2 of 2 

#> $`2837007`
#> $`2837007`$name
#> [1] "Marcus Beck"
#> 
#> $`2837007`$location
#> [1] "Irvine, California"
#> 
#> $`2837007`$units
#> [1] "mi" "ft"
#> 
#> $`2837007`$monthly
#>        month     miles hours elev_gain
#> 1 2019-05-01 351.16172    25      1266
#> 2 2019-06-01 283.40678    21      1132
#> 3 2019-07-01 273.54686    20       893
#> 4 2019-08-01 342.39727    25      1126
#> 5 2019-09-01 284.57744    20      1041
#> 6 2019-10-01  48.97586     7       773
#> 7 2019-11-01  56.55658     4       147
#> 
#> $`2837007`$recent
#>           id           name type startDateLocal distance elevation movingTime
#> 1 2859540960   Morning Ride ride     2019-11-11      3.5        18      16:29
#> 2 2858493438 Afternoon Ride ride     2019-11-11     36.2       339    2:21:55
#> 3 2858170842    Morning Run  run     2019-11-08      3.2        47      23:36
#> 
#> $`2837007`$achievements
#>                          description             timeago
#> 1   2nd best estimated 1 mile effort 2019-11-09 12:01:09
#> 2       3rd best estimated 1k effort 2019-11-09 12:01:09
#> 3 3rd best estimated 1/2 mile effort 2019-11-09 12:01:09
#> 4     3rd best estimated 400m effort 2019-11-09 12:01:09
#> 5  2nd fastest time on 21-9st Sprint 2019-11-12 12:10:31
#> 
#> 
#> $`2527465`
#> $`2527465`$name
#> [1] "Pedro Villarroel"
#> 
#> $`2527465`$location
#> [1] "Buenos Aires, Ciudad Autónoma de Buenos Aires, Argentina"
#> 
#> $`2527465`$units
#> [1] "km" "m" 
#> 
#> $`2527465`$monthly
#>        month   miles hours elev_gain
#> 1 2019-05-01 593.138    30      5881
#> 2 2019-06-01 335.506    18      3265
#> 3 2019-07-01  95.858     5      1057
#> 4 2019-08-01   6.859     0        54
#> 5 2019-09-01   0.000     0         0
#> 6 2019-10-01  76.962     4       877
#> 7 2019-11-01  53.719     3       800
#> 
#> $`2527465`$recent
#>           id                                name type startDateLocal distance
#> 1 2853265134                 Training primaveral ride     2019-11-09     28.2
#> 2 2835378038   Comprando el desayuno en La Perla ride     2019-11-02     11.4
#> 3 2834222745 Vitius + Python Programming en EANT ride     2019-11-01     14.0
#>   elevation movingTime
#> 1       502    1:44:22
#> 2       118      38:05
#> 3       181      49:04
#> 
#> $`2527465`$achievements
#>                                                            description
#> 1                                      PR on Retiro-Olleros (por Lib.)
#> 2                             PR on Retiro hasta Au Illia x Libertador
#> 3                                            PR on Corrientes a Retiro
#> 4             2nd fastest time on La Pampa - Juramento x Virrey Vértiz
#> 5 2nd fastest time on Túnel Libertador hasta Barrancas x Virrey Vértiz
#>               timeago
#> 1 2019-11-09 17:09:17
#> 2 2019-11-09 17:09:17
#> 3 2019-11-09 17:09:17
#> 4 2019-11-09 17:09:17
#> 5 2019-11-09 17:09:17
#> 
#>