Get data for an athlete

Get data for an athlete by web scraping, does not require authentication.

athl_fun(athl_num, trace = TRUE)

Arguments

athl_num
numeric vector of athlete id(s) used by Strava

trace
logical indicating if output is returned to console

Value

A list for each athlete, where each element is an additional list with elements for the units of measurement, location, current month data, monthly data, year-to-date data, and an all-time summary. The list elements are named using the athlete id numbers. NA will be returned if the data for an athlete could not be accessed.

Details

The athlete id is assigned to the user during registration with Strava and this must be known to use the function. Some users may have privacy settings that prevent public access to account information (NA will be returned). The function scrapes data from https://www.strava.com/athletes/ with the appended athlete id, e.g., https://www.strava.com/athletes/2837007. Opening the URL in a web browser can verify if the data can be scraped. Logging in to the Strava account on the website may also be required before using this function.

Examples

## single athlete
athl_fun(2837007)

#> 2837007 1 of 1 

#> $`2837007`
#> $`2837007`$name
#> [1] "Marcus Beck"
#> 
#> $`2837007`$location
#> [1] "Irvine, California"
#> 
#> $`2837007`$units
#> [1] "mi" "ft"
#> 
#> $`2837007`$monthly
#>        month     miles hours elev_gain
#> 1 2019-02-01 247.53005    18       679
#> 2 2019-03-01 343.69594    24      1039
#> 3 2019-04-01 332.47584    23       990
#> 4 2019-05-01 351.16172    25      1266
#> 5 2019-06-01 283.40678    21      1132
#> 6 2019-07-01 273.54686    20       893
#> 7 2019-08-01  24.84863     1        75
#> 
#> $`2837007`$recent
#>           id           name type startDateLocal distance elevation movingTime
#> 1 2584096258    Morning Run  run     2019-08-01      3.2        72      25:52
#> 2 2583013926 Afternoon Ride ride     2019-08-01     10.9       126      41:55
#> 3 2581426804   Morning Ride ride     2019-07-31     10.6        51      40:06
#> 
#> $`2837007`$achievements
#>                           description             timeago
#> 1       2nd fastest time on Park Push 2019-07-27 00:41:33
#> 2    2nd fastest time on Bryan bridge 2019-07-24 00:44:12
#> 3 2nd fastest time on Pitchin' A Tent 2019-07-18 00:42:44
#> 
#> 

## multiple athletes
athl_fun(c(2837007, 2527465))

#> 2837007 1 of 2 
#> 2527465 2 of 2 

#> $`2837007`
#> $`2837007`$name
#> [1] "Marcus Beck"
#> 
#> $`2837007`$location
#> [1] "Irvine, California"
#> 
#> $`2837007`$units
#> [1] "mi" "ft"
#> 
#> $`2837007`$monthly
#>        month     miles hours elev_gain
#> 1 2019-02-01 247.53005    18       679
#> 2 2019-03-01 343.69594    24      1039
#> 3 2019-04-01 332.47584    23       990
#> 4 2019-05-01 351.16172    25      1266
#> 5 2019-06-01 283.40678    21      1132
#> 6 2019-07-01 273.54686    20       893
#> 7 2019-08-01  24.84863     1        75
#> 
#> $`2837007`$recent
#>           id           name type startDateLocal distance elevation movingTime
#> 1 2584096258    Morning Run  run     2019-08-01      3.2        72      25:52
#> 2 2583013926 Afternoon Ride ride     2019-08-01     10.9       126      41:55
#> 3 2581426804   Morning Ride ride     2019-07-31     10.6        51      40:06
#> 
#> $`2837007`$achievements
#>                           description             timeago
#> 1       2nd fastest time on Park Push 2019-07-27 00:41:33
#> 2    2nd fastest time on Bryan bridge 2019-07-24 00:44:12
#> 3 2nd fastest time on Pitchin' A Tent 2019-07-18 00:42:44
#> 
#> 
#> $`2527465`
#> $`2527465`$name
#> [1] "Pedro Villarroel"
#> 
#> $`2527465`$location
#> [1] "Buenos Aires, Ciudad Autónoma de Buenos Aires, Argentina"
#> 
#> $`2527465`$units
#> [1] "km" "m" 
#> 
#> $`2527465`$monthly
#>        month   miles hours elev_gain
#> 1 2019-02-01 258.484    14      3013
#> 2 2019-03-01 387.058    23      4521
#> 3 2019-04-01 361.513    18      3578
#> 4 2019-05-01 593.138    30      5881
#> 5 2019-06-01 335.506    18      3265
#> 6 2019-07-01  95.858     5      1057
#> 7 2019-08-01   0.000     0         0
#> 
#> $`2527465`$recent
#>           id   name type startDateLocal distance elevation movingTime
#> 1 2554963755 Vitius ride     2019-07-22     11.3       129      40:28
#> 2 2543953790 Vitius ride     2019-07-18     10.8       119      35:16
#> 3 2541352625 Vitius ride     2019-07-17     10.7       119      34:48
#> 
#> $`2527465`$achievements
#> list()
#> 
#>