Get data for an athlete

Get data for an athlete by web scraping, does not require authentication.

athl_fun(athl_num, trace = TRUE)

Arguments

athl_num
numeric vector of athlete id(s) used by Strava

trace
logical indicating if output is returned to console

Value

A list for each athlete, where each element is an additional list with elements for the units of measurement, location, current month data, monthly data, year-to-date data, and an all-time summary. The list elements are named using the athlete id numbers. NA will be returned if the data for an athlete could not be accessed.

Details

The athlete id is assigned to the user during registration with Strava and this must be known to use the function. Some users may have privacy settings that prevent public access to account information (NA will be returned). The function scrapes data from https://www.strava.com/athletes/ with the appended athlete id, e.g., https://www.strava.com/athletes/2837007. Opening the URL in a web browser can verify if the data can be scraped. Logging in to the Strava account on the website may also be required before using this function.

Examples

## single athlete
athl_fun(2837007)

#> 2837007 1 of 1 

#> $`2837007`
#> $`2837007`$name
#> [1] "Marcus Beck"
#> 
#> $`2837007`$location
#> [1] "Irvine, California"
#> 
#> $`2837007`$units
#> [1] "mi" "ft"
#> 
#> $`2837007`$monthly
#>        month     miles hours elev_gain
#> 1 2018-12-01 161.58634    14       964
#> 2 2019-01-01 391.10905    28      1110
#> 3 2019-02-01 247.53005    18       679
#> 4 2019-03-01 343.69594    24      1039
#> 5 2019-04-01 332.47584    23       990
#> 6 2019-05-01 351.16172    25      1266
#> 7 2019-06-01  15.70267     1        83
#> 
#> $`2837007`$recent
#>           id           name type startDateLocal distance elevation movingTime
#> 1 2420187776   Morning Ride ride     2019-06-02     10.6        51      39:40
#> 2 2419088213  Afternoon Run  run     2019-06-02      5.0       223      36:55
#> 3 2419088214 Afternoon Ride ride     2019-05-31     10.8       117      41:44
#> 
#> $`2837007`$achievements
#>                      description             timeago
#> 1 2nd best estimated 400m effort 2019-05-22 22:43:09
#> 2           PR on Trailwood down 2019-06-02 23:19:02
#> 3   PR on State Street Ramp Down 2019-05-23 13:36:00
#> 4      PR on Capital Hill Repeat 2019-05-23 13:36:00
#> 5        PR on State Street Ramp 2019-05-23 13:36:00
#> 
#> 

## multiple athletes
athl_fun(c(2837007, 2527465))

#> 2837007 1 of 2 
#> 2527465 2 of 2 

#> $`2837007`
#> $`2837007`$name
#> [1] "Marcus Beck"
#> 
#> $`2837007`$location
#> [1] "Irvine, California"
#> 
#> $`2837007`$units
#> [1] "mi" "ft"
#> 
#> $`2837007`$monthly
#>        month     miles hours elev_gain
#> 1 2018-12-01 161.58634    14       964
#> 2 2019-01-01 391.10905    28      1110
#> 3 2019-02-01 247.53005    18       679
#> 4 2019-03-01 343.69594    24      1039
#> 5 2019-04-01 332.47584    23       990
#> 6 2019-05-01 351.16172    25      1266
#> 7 2019-06-01  15.70267     1        83
#> 
#> $`2837007`$recent
#>           id           name type startDateLocal distance elevation movingTime
#> 1 2420187776   Morning Ride ride     2019-06-02     10.6        51      39:40
#> 2 2419088213  Afternoon Run  run     2019-06-02      5.0       223      36:55
#> 3 2419088214 Afternoon Ride ride     2019-05-31     10.8       117      41:44
#> 
#> $`2837007`$achievements
#>                      description             timeago
#> 1 2nd best estimated 400m effort 2019-05-22 22:43:09
#> 2           PR on Trailwood down 2019-06-02 23:19:02
#> 3   PR on State Street Ramp Down 2019-05-23 13:36:00
#> 4      PR on Capital Hill Repeat 2019-05-23 13:36:00
#> 5        PR on State Street Ramp 2019-05-23 13:36:00
#> 
#> 
#> $`2527465`
#> $`2527465`$name
#> [1] "Pedro Villarroel"
#> 
#> $`2527465`$location
#> [1] "Buenos Aires, Ciudad Autónoma de Buenos Aires, Argentina"
#> 
#> $`2527465`$units
#> [1] "km" "m" 
#> 
#> $`2527465`$monthly
#>        month   miles hours elev_gain
#> 1 2018-12-01 257.459    15      3370
#> 2 2019-01-01 210.299    10      2560
#> 3 2019-02-01 258.484    14      3013
#> 4 2019-03-01 387.058    23      4521
#> 5 2019-04-01 361.513    18      3578
#> 6 2019-05-01 593.138    30      5881
#> 7 2019-06-01  39.722     1       525
#> 
#> $`2527465`$recent
#>           id                       name type startDateLocal distance elevation
#> 1 2415604153            48x15 Test ride ride     2019-06-01     39.7       526
#> 2 2412950301 Back to 48x15 <U+0001F625> ride     2019-05-31     16.8       141
#> 3 2409000032                  Windy KDT ride     2019-05-30     36.2       246
#>   movingTime
#> 1    1:59:02
#> 2    1:27:55
#> 3    1:33:39
#> 
#> $`2527465`$achievements
#>                                              description             timeago
#> 1                            PR on subida Thompson-Craig 2019-06-01 18:40:56
#> 2           PR on Honorio: de J B Justo al Cid Campeador 2019-06-01 18:40:56
#> 3                               PR on Honorio Pueyrredon 2019-06-01 18:40:56
#> 4 PR on Warnes Honorio: de la Vía San Martín a J B Justo 2019-06-01 18:40:56
#> 5                    PR on Trepada Puente Av. San Martin 2019-06-01 18:40:56
#> 
#>