converts a list of activities into a dataframe

converts a list of activities into a dataframe

compile_activities(actlist, acts = NULL, units = "metric")

Arguments

actlist
an activities list returned by get_activity_list

acts
numeric indicating which activities to compile starting with most recent, defaults to all

units
chr string indicating metric or imperial

...
arguments passed to or from other methods

Value

An actitities frame object (actframe that includes a data frame for the data and attributes for the distance, speed, and elevation units

Details

each activity has a value for every column present across all activities, with NAs populating empty values

Examples

 

# NOT RUN {
stoken <- httr::config(token = strava_oauth(app_name, app_client_id, app_secret, cache = TRUE))

my_acts <- get_activity_list(stoken)

acts_data <- compile_activities(my_acts)

# show attributes
attr(acts_data, 'unit_type')
attr(acts_data, 'unit_vals')
# }