Convert a set of streams of a single activity into a dataframe

Convert a set of streams of a single activity into a dataframe, with the retrieved columns.

compile_activity_streams(streams, id = NULL)

Arguments

streams
a list containing details of the Strava streams of a single activity (output of get_streams)

id
if not missing, the activity id of the stream (will be appended to the data.frame, if non-empty)

Value

data frame where every column is the stream data for the retrieved types.

Details

used internally in get_activity_streams

Examples

if (FALSE) {
stoken <- httr::config(token = strava_oauth(app_name, app_client_id, app_secret, cache = TRUE))

act_id <- 351217692
streams <- get_streams(stoken, id = act_id, types = list('distance', 'latlng'))

compile_activity_streams(streams, id = act_id)}