Generata Strava API authentication token

Generate a token for the user and the desired scope. The user is sent to the strava authentication page if he/she hasn't given permission to the app yet, else, is sent to the app webpage.

strava_oauth(app_name, app_client_id, app_secret, app_scope = "public",
  cache = FALSE)

Arguments

app_name
chr string for name of the app

app_client_id
chr string for ID received when the app was registered

app_secret
chr string for secret received when the app was registered

app_scope
chr string for scope of authentication, Must be "public", "write", "view_private", or "view_private,write"

cache
logical to cache the token

Details

The app_name, app_client_id, and app_secret are specific to the user and can be obtained by registering an app on the Strava API authentication page: http://strava.github.io/api/v3/oauth/. This requires a personal Strava account.